+++ backend/src/handlers/localgame.handler.ts
@@ -6,7 +6,7 @@ import localGameService from '../services/localgame.service';
 const localGameHandler = () => {
   const createLocalGame = async (req: FastifyRequest, rep: FastifyReply) => {
     try {
-      await localGameService.createLocalGame(req.body, req.user.userId);
+      await localGameService.createLocalGame(req.body, req.user.userId); // req.body에서 오류발생
       handlerUtil.handleSuccess(rep, SUCCESS_MESSAGE.registerOK);
     } catch (error) {
       handlerUtil.handleError(rep, ERROR_MESSAGE.badRequest, error);
